new architecure for usage reports
