+++ server/server.go
@@ -33,10 +33,10 @@ import (
 	_ "github.com/trackit/trackit-server/costs/anomalies"
 	_ "github.com/trackit/trackit-server/costs/diff"
 	_ "github.com/trackit/trackit-server/costs/tags"
-	_ "github.com/trackit/trackit-server/ec2"
+	_ "github.com/trackit/trackit-server/usageReports/ec2"
 	"github.com/trackit/trackit-server/periodic"
 	_ "github.com/trackit/trackit-server/plugins"
-	_ "github.com/trackit/trackit-server/rds"
+	_ "github.com/trackit/trackit-server/usageReports/rds"
 	_ "github.com/trackit/trackit-server/reports"
 	"github.com/trackit/trackit-server/routes"
 	_ "github.com/trackit/trackit-server/s3/costs"
