+++ usageReports/rds/prepare_response.go
@@ -1,4 +1,4 @@
-//   Copyright 2018 MSolution.IO
+//   Copyright 2017 MSolution.IO
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-package ec2
+package rds
 
 import (
 	"strings"
@@ -21,6 +21,8 @@ import (
 	"encoding/json"
 
 	"gopkg.in/olivere/elastic.v5"
+
+	"github.com/trackit/trackit-server/aws/usageReports/rds"
 )
 
 type (
@@ -42,68 +44,37 @@ type (
 		} `json:"accounts"`
 	}
 
-	// Structure that allow to parse ES response for EC2 usage report
-	ResponseEc2 struct {
+	// Structure that allow to parse ES response for RDS usage report
+	ResponseRds struct {
 		TopReports struct {
 			Buckets []struct {
 				TopReportsHits struct {
 					Hits struct {
 						Hits []struct {
-							Source Report `json:"_source"`
+							Source rds.Report `json:"_source"`
 						} `json:"hits"`
 					} `json:"hits"`
 				} `json:"top_reports_hits"`
 			} `json:"buckets"`
 		} `json:"top_reports"`
 	}
-
-	// Report format for EC2 usage
-	Report struct {
-		Account    string     `json:"account"`
-		ReportDate string     `json:"reportDate"`
-		ReportType string     `json:"reportType"`
-		Instances  []Instance `json:"instances"`
-	}
-
-	// Instance contains stats of an EC2 instance
-	Instance struct {
-		Id         string             `json:"id"`
-		Region     string             `json:"region"`
-		State      string             `json:"state"`
-		Purchasing string             `json:"purchasing"`
-		Cost       float64            `json:"cost"`
-		CpuAverage float64            `json:"cpuAverage"`
-		CpuPeak    float64            `json:"cpuPeak"`
-		NetworkIn  int64              `json:"networkIn"`
-		NetworkOut int64              `json:"networkOut"`
-		IORead     map[string]float64 `json:"ioRead"`
-		IOWrite    map[string]float64 `json:"ioWrite"`
-		KeyPair    string             `json:"keyPair"`
-		Type       string             `json:"type"`
-		Tags       map[string]string  `json:"tags"`
-	}
 )
 
 // addCostToReport adds cost for each instance based on billing data
-func addCostToReport(report Report, costs ResponseCost) (Report) {
+func addCostToReport(report rds.Report, costs ResponseCost) (rds.Report) {
 	for _, accounts := range costs.Accounts.Buckets {
 		if accounts.Key != report.Account {
 			continue
 		}
 		for _, instance := range accounts.Instances.Buckets {
 			for i := range report.Instances {
-				if strings.Contains(instance.Key, report.Instances[i].Id) {
+				if report.Instances[i].CostDetail == nil {
+					report.Instances[i].CostDetail = make(map[string]float64, 0)
+				}
+				split := strings.Split(instance.Key, ":")
+				if len(split) == 7 && split[6] == report.Instances[i].DBInstanceIdentifier {
 					report.Instances[i].Cost += instance.Cost.Value
-				}
-				for volume := range report.Instances[i].IOWrite {
-					if volume == instance.Key {
-						report.Instances[i].Cost += instance.Cost.Value
-					}
-				}
-				for volume := range report.Instances[i].IORead {
-					if volume == instance.Key {
-						report.Instances[i].Cost += instance.Cost.Value
-					}
+					report.Instances[i].CostDetail["instance"] += instance.Cost.Value
 				}
 			}
 		}
@@ -111,19 +82,19 @@ func addCostToReport(report Report, costs ResponseCost) (Report) {
 	return report
 }
 
-// prepareResponseEc2 parses the results from elasticsearch and returns the EC2 usage report
-func prepareResponseEc2(ctx context.Context, resEc2 *elastic.SearchResult, resCost *elastic.SearchResult) ([]Report, error) {
-	ec2  := ResponseEc2{}
+// prepareResponseRdsDaily parses the results from elasticsearch and returns the RDS report
+func prepareResponseRdsDaily(ctx context.Context, resRds *elastic.SearchResult, resCost *elastic.SearchResult) ([]rds.Report, error) {
+	parsedRds  := ResponseRds{}
 	cost := ResponseCost{}
-	reports := []Report{}
-	err := json.Unmarshal(*resEc2.Aggregations["top_reports"], &ec2.TopReports)
+	reports := []rds.Report{}
+	err := json.Unmarshal(*resRds.Aggregations["top_reports"], &parsedRds.TopReports)
 	if err != nil {
 		return nil, err
 	}
 	if resCost != nil {
 		json.Unmarshal(*resCost.Aggregations["accounts"], &cost.Accounts)
 	}
-	for _, account := range ec2.TopReports.Buckets {
+	for _, account := range parsedRds.TopReports.Buckets {
 		if len(account.TopReportsHits.Hits.Hits) > 0 {
 			report := account.TopReportsHits.Hits.Hits[0].Source
 			report = addCostToReport(report, cost)
@@ -133,11 +104,11 @@ func prepareResponseEc2(ctx context.Context, resEc2 *elastic.SearchResult, resCo
 	return reports, nil
 }
 
-// prepareResponseEc2Monthly parses the results from elasticsearch and returns the EC2 usage report
-func prepareResponseEc2Monthly(ctx context.Context, resEc2 *elastic.SearchResult) ([]Report, error) {
-	response := ResponseEc2{}
-	reports := []Report{}
-	err := json.Unmarshal(*resEc2.Aggregations["top_reports"], &response.TopReports)
+// prepareResponseRdsMonthly parses the results from elasticsearch and returns the RDS usage report
+func prepareResponseRdsMonthly(ctx context.Context, resRds *elastic.SearchResult) ([]rds.Report, error) {
+	response := ResponseRds{}
+	reports := []rds.Report{}
+	err := json.Unmarshal(*resRds.Aggregations["top_reports"], &response.TopReports)
 	if err != nil {
 		return nil, err
 	}
@@ -149,7 +120,7 @@ func prepareResponseEc2Monthly(ctx context.Context, resEc2 *elastic.SearchResult
 	return reports, nil
 }
 
-func isInstanceUnused(instance Instance) bool {
+func isInstanceUnused(instance rds.Instance) bool {
 	average := instance.CpuAverage
 	peak    := instance.CpuPeak
 	if peak >= 60.0 {
@@ -160,9 +131,9 @@ func isInstanceUnused(instance Instance) bool {
 	return true
 }
 
-// prepareResponseEc2Unused filter reports to get the top instances
-func prepareResponseEc2Unused(params ec2UnusedQueryParams, reports []Report) (int, []Instance, error) {
-	instances := []Instance{}
+// prepareResponseRdsUnused filter reports to get the top instances
+func prepareResponseRdsUnused(params rdsUnusedQueryParams, reports []rds.Report) (int, []rds.Instance, error) {
+	instances := []rds.Instance{}
 	for _, report := range reports {
 		for _, instance := range report.Instances {
 			if isInstanceUnused(instance) {
